+++ HabarBankAPI.Application/Interfaces/ISendingService.cs
@@ -12,9 +12,9 @@ namespace HabarBankAPI.Application.Interfaces
     public interface ISendingService : IAppService
     {
         public Task CreateTransfer(SendingDTO sendingDTO);
-        public Task<SendingDTO> GetTransferByTransferId(int sendingId);
-        public Task<IList<SendingDTO>> GetTransfersBySubstanceId(int substanceId);
-        public Task<IList<SendingDTO>> GetTransfersByUserId(int userId);
-        public Task SetTransferStatus(int sendingId, bool sendingEnabled);
+        public Task<SendingDTO> GetTransferByTransferId(long sendingId);
+        public Task<IList<SendingDTO>> GetTransfersBySubstanceId(long substanceId);
+        public Task<IList<SendingDTO>> GetTransfersByUserId(long userId);
+        public Task SetTransferStatus(long sendingId, bool sendingEnabled);
     }
 }
